# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BIRU MENARA API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /team-members:
    get:
      operationId: listTeamMembers
      tags: [teamMembers]
      summary: List all team members
      responses:
        "200":
          description: List of team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: createTeamMember
      tags: [teamMembers]
      summary: Create team member (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /team-members/{id}:
    get:
      operationId: getTeamMember
      tags: [teamMembers]
      summary: Get team member by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
        "404":
          description: Not found
    put:
      operationId: updateTeamMember
      tags: [teamMembers]
      summary: Update team member (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
        "404":
          description: Not found
    delete:
      operationId: deleteTeamMember
      tags: [teamMembers]
      summary: Delete team member (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ---- Downloads ----
...
components:
  schemas:
...
    TeamMember:
      type: object
      properties:
        id:
          type: integer
        role:
          type: string
        name:
          type: string
        memberType:
          type: string
        sortOrder:
          type: integer
        imageUrl:
          type: ["string", "null"]
      required: [id, role, name, memberType, sortOrder]
    TeamMemberInput:
      type: object
      properties:
        role:
          type: string
        name:
          type: string
        memberType:
          type: string
        sortOrder:
          type: integer
        imageUrl:
          type: ["string", "null"]
      required: [role, name, memberType]
    TeamMemberUpdate:
      type: object
      properties:
        role:
          type: string
        name:
          type: string
        memberType:
          type: string
        sortOrder:
          type: integer
        imageUrl:
          type: ["string", "null"]
...