# lib/api-zod/src/generated/types/teamMemberUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * BIRU MENARA API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TeamMemberUpdate {
  role?: string;
  name?: string;
  memberType?: string;
  sortOrder?: number;
  /** @nullable */
  imageUrl?: string | null;
}